split into mutiple files define output file improve input/output methods
