+++ graph/graph.h
@@ -3,8 +3,6 @@
 #include <unordered_set>
 #include <climits>
 #include <vector>
-#include <fstream>
-using namespace std;
 
 class Graph{
 private:
@@ -146,123 +144,3 @@ public:
         return serialize(dist, INF);
     }
 };
-
-void printMat(const vector<vector<int> > &v) {
-    int n = v.size();
-    cout<<endl;
-    for(int i=0; i<n; i++){
-        for(int j=0; j<n; j++){
-            cout<<v[i][j]<<'\t';
-        }
-        cout<<endl;
-    }
-    cout<<endl;
-}
-
-void printVec(const vector<int> &v) {
-    for(int x: v){
-        cout<<x<<'\t';
-    }
-    cout<<endl;
-}
-
-class SolutionSubRoutine {
-public:
-    virtual void precompute(const Graph &graph) {
-        // do nothing
-    }
-    virtual vector<int> ditancesFromSource(int source, const Graph &graph) {
-        return vector<int>();
-    }
-};
-
-class BellmanFordSubRoutine: public SolutionSubRoutine {
-public:
-    BellmanFordSubRoutine() {}
-    vector<int> ditancesFromSource(int source, const Graph &graph) override {
-        return graph.bellmanFord(source);
-    }
-};
-
-class FloyedWarshallSubRoutine: public SolutionSubRoutine {
-private:
-    vector<vector<int> > v;
-public:
-    FloyedWarshallSubRoutine() {}
-    void precompute(const Graph &graph) override {
-        v = graph.floydWarshall();
-    }
-    vector<int> ditancesFromSource(int source, const Graph &graph) override {
-        return v[source];
-    }
-};
-
-vector<int> fImpl(vector<vector<int> > &v, SolutionSubRoutine *routine) {
-    int n = 1+v.size();
-    Graph graph = new Graph(true, 1000000); // assuming INFINITY = 10^6
-    for(auto p: v){
-        graph.addEdge(p[0],p[1],2 - (p[1]%2));
-        graph.addEdge(p[1],p[0],2 - (p[0]%2));
-    }
-    routine->precompute(graph);
-    
-    vector<int> ans(n,0);
-    vector<int> temp;
-    for(int i=0; i<n; i++){
-        temp = routine->ditancesFromSource(i, graph);
-        for(int x: temp){
-            ans[i] = max(ans[i], x);
-        }
-    }
-    return ans;
-}
-
-vector<int> f(vector<vector<int> > &v) {
-    // return fImpl(v, new BellmanFordSubRoutine());
-    return fImpl(v, new FloyedWarshallSubRoutine());
-}
-
-
-template <class T>
-T inputImpl(ifstream &fs, T*){
-    T x;
-    fs>>x;
-    return x;
-}
-
-template <class X>
-vector<X> inputImpl(ifstream &fs, vector<X> *){
-    int n;
-    fs>>n;
-    vector<X> v;
-    while(n--){
-        v.push_back(inputImpl(fs, static_cast<X*>(0)));
-    }
-    return v;
-}
-
-template <class T>
-T input(ifstream &fs){
-    return inputImpl(fs, static_cast<T*>(0));
-}
-
-int main(int argc, char* argv[]) {
-    string relativePath = "";
-    if(argc>1){
-        relativePath = argv[1];
-    }
-    string inputFilePath = relativePath + "/input.in";
-    ifstream inputFile;
-    inputFile.open(inputFilePath);  // relative to the build location
-    vector<vector<vector<int> > > inputs = input<vector<vector<vector<int> > > >(inputFile);
-    vector<int> output;
-    for(auto &v: inputs){
-        output = f(v);
-        printVec(output);
-    }
-    return 0;
-}
-
-// implement pending methods, bellmon-ford, dijkstras, johnsons etc.
-// run it for the leetcode question
-// improve the INF handling by using unordered_map instead of vector
