+++ graph/output.out
@@ -1,0 +1,3 @@
+2	4	3	
+1	2	
+4	6	3	5	5
